Different Employers are now able to tie tip rates to their current and future employees. Now for cleaning and testing.

+++ src/main/java/com/tipout/Tipout/controllers/AuthenticationController.java
@@ -24,7 +24,7 @@ public class AuthenticationController {
     private static final String userSessionKey = "employer";
 
     public Employer getEmployerFromSession(HttpSession session) {
-        Integer employerId = (Integer) session.getAttribute(userSessionKey);
+        Long employerId = (Long) session.getAttribute(userSessionKey);
         if (employerId == null) {
             return null;
         }
@@ -43,7 +43,7 @@ public class AuthenticationController {
     }
 
     public boolean inSession(HttpSession session){
-        Integer employerId = (Integer) session.getAttribute(userSessionKey);
+        Long employerId = (Long) session.getAttribute(userSessionKey);
         if (employerId == null) {
             return false;
         }
