+++ src/main/java/com/tipout/Tipout/models/data/BOHRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BOHRepository extends CrudRepository<BOH,Integer> {
+public interface BOHRepository extends CrudRepository<BOH,Long> {
 }
