+++ src/main/java/com/tipout/Tipout/models/data/BartenderRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BartenderRepository extends CrudRepository<Bartender,Integer> {
+public interface BartenderRepository extends CrudRepository<Bartender,Long> {
 }
