+++ src/main/java/com/tipout/Tipout/models/data/BusserRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BusserRepository extends CrudRepository<Busser,Integer> {
+public interface BusserRepository extends CrudRepository<Busser,Long> {
 }
