+++ src/main/java/com/tipout/Tipout/models/data/EmployeeRepository.java
@@ -11,25 +11,25 @@ import java.util.List;
 
 
 @Repository
-public interface EmployeeRepository extends CrudRepository<Employee,Integer>{
+public interface EmployeeRepository extends CrudRepository<Employee,Long>{
     @Transactional
     @Modifying
     @Query(value="DElETE FROM tipout.TipsCollected_employeeTipsMap Where employeeTipsMap_KEY = ?1", nativeQuery = true)
-    void completelyDeleteEmployeeTipRecord(Integer id);
+    void completelyDeleteEmployeeTipRecord(Long id);
     @Transactional
     @Modifying
     @Query(value="DELETE FROM tipscollected_moneyhandlertipsmap Where moneyHandlerTipsMap_KEY = ?1", nativeQuery = true)
-    void completelyDeleteMoneyhandlerTipRecord(Integer id);
+    void completelyDeleteMoneyhandlerTipRecord(Long id);
 
     @Transactional
     @Modifying
     @Query(value="DELETE FROM tipscollected_nonmoneyhandlertipsmap Where nonmoneyHandlerTipsMap_KEY = ?1", nativeQuery = true)
-    void completelyDeleteNonMoneyhandlerTipRecord(Integer id);
+    void completelyDeleteNonMoneyhandlerTipRecord(long id);
 
     @Query(value="SELECT *, 0 AS clazz_ FROM tipout.EMPLOYEE Where deleted = false AND employer_id = ?1", nativeQuery = true)
-    List<Employee> findCurrentEmployees(Integer id);
+    List<Employee> findCurrentEmployees(long id);
 
     @Query(value="SELECT *, 0 AS clazz_ FROM tipout.EMPLOYEE Where deleted = true AND employer_id = ?1", nativeQuery = true)
-    List<Employee> findArhievedEmployees(Integer id);
+    List<Employee> findArhievedEmployees(long id);
 
 }
