+++ src/main/java/com/tipout/Tipout/models/data/EmployerRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmployerRepository extends CrudRepository<Employer, Integer> {
+public interface EmployerRepository extends CrudRepository<Employer, Long> {
     Employer findByUsername(String username);
 }
