+++ src/main/java/com/tipout/Tipout/models/data/MoneyHandlerRepository.java
@@ -1,8 +1,6 @@
 package com.tipout.Tipout.models.data;
 
-import com.tipout.Tipout.models.Employee;
 import com.tipout.Tipout.models.Employees.MoneyHandler;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,6 +8,6 @@ import java.util.List;
 
 
 @Repository
-public interface MoneyHandlerRepository extends CrudRepository<MoneyHandler,Integer>{
-    List<MoneyHandler> findAllByDeletedFalseAndEmployer_Id(Integer Employer_Id);
+public interface MoneyHandlerRepository extends CrudRepository<MoneyHandler,Long>{
+    List<MoneyHandler> findAllByDeletedFalseAndEmployer_Id(long Employer_Id);
 }
