+++ src/main/java/com/tipout/Tipout/models/data/NonMoneyHandlerRepository.java
@@ -1,8 +1,6 @@
 package com.tipout.Tipout.models.data;
 
-import com.tipout.Tipout.models.Employee;
 import com.tipout.Tipout.models.Employees.NonMoneyHandler;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,6 +8,6 @@ import java.util.List;
 
 
 @Repository
-public interface NonMoneyHandlerRepository extends CrudRepository<NonMoneyHandler,Integer>{
-    List<NonMoneyHandler> findAllByDeletedFalseAndEmployer_Id(Integer Employer_Id);
+public interface NonMoneyHandlerRepository extends CrudRepository<NonMoneyHandler,Long>{
+    List<NonMoneyHandler> findAllByDeletedFalseAndEmployer_Id(long Employer_Id);
 }
